+++ src/apps/Permissions/EntitySelector.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { DropDown, Field, TextInput } from '@aragon/ui'
-import { getAnyAddress, getEmptyAddress } from '../../aragonos-utils'
+import { getAnyAddress, getEmptyAddress } from '../../web3-utils'
 import AppInstanceLabel from './AppInstanceLabel'
 
 class EntitySelector extends React.Component {
