+++ src/apps/Permissions/ManageRolePanel.js
@@ -2,8 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { SidePanel, DropDown, Info, Field, Button, TextInput } from '@aragon/ui'
 import { PermissionsConsumer } from '../../contexts/PermissionsContext'
-import { isEmptyAddress } from '../../aragonos-utils'
-import { isAddress } from '../../web3-utils'
+import { isAddress, isEmptyAddress } from '../../web3-utils'
 import AppInstanceLabel from './AppInstanceLabel'
 import IdentityBadge from '../../components/IdentityBadge'
 import EntitySelector from './EntitySelector'
